added BaseTest constants, utils and models moved to src.main.java (steps cant be moved to main because its ude a ResAssured) replaced the report with a new one
