+++ src/main/java/ru/praktikum/services/qa/scooter/constants/Colors.java
