+++ src/main/java/ru/praktikum/services/qa/scooter/constants/Messages.java
