+++ src/main/java/ru/praktikum/services/qa/scooter/models/Courier.java
