+++ src/main/java/ru/praktikum/services/qa/scooter/models/Order.java
