+++ src/main/java/ru/praktikum/services/qa/scooter/utils/Utils.java
