+++ src/test/java/ru/praktikum/services/qa/scooter/tests/CourierCreatePositiveTest.java
@@ -1,4 +1,4 @@
-package ru.praktikum.services.qa.scooter;
+package ru.praktikum.services.qa.scooter.tests;
 
 import io.qameta.allure.Description;
 import io.qameta.allure.junit4.DisplayName;
@@ -7,6 +7,7 @@ import io.restassured.response.Response;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
+import ru.praktikum.services.qa.scooter.BaseTest;
 import ru.praktikum.services.qa.scooter.models.Courier;
 
 import static ru.praktikum.services.qa.scooter.constants.Messages.*;
@@ -14,16 +15,12 @@ import static ru.praktikum.services.qa.scooter.constants.Uri.QA_SQOOTER_PRACTICU
 import static ru.praktikum.services.qa.scooter.steps.StepsCourier.*;
 import static ru.praktikum.services.qa.scooter.utils.Utils.randomString;
 
-public class CourierCreatePositiveTest {
+public class CourierCreatePositiveTest extends BaseTest {
 
     private static final String RANDOM_LOGIN_OF_COURIER = randomString();
     private static final String RANDOM_PASSWORD_OF_COURIER = randomString();
     Courier courier;
 
-    @Before
-    public void setUp() {
-        RestAssured.baseURI= QA_SQOOTER_PRACTICUM_PROD;
-    }
 
     @After
     public void tearDown() {
