+++ src/test/java/ru/praktikum/services/qa/scooter/tests/OrdersGetTest.java
@@ -1,4 +1,4 @@
-package ru.praktikum.services.qa.scooter;
+package ru.praktikum.services.qa.scooter.tests;
 
 import io.qameta.allure.Description;
 import io.qameta.allure.junit4.DisplayName;
@@ -6,22 +6,15 @@ import io.restassured.RestAssured;
 import io.restassured.response.Response;
 import org.junit.Before;
 import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.junit.runners.Parameterized;
+import ru.praktikum.services.qa.scooter.BaseTest;
 import ru.praktikum.services.qa.scooter.models.Order;
 
-import java.util.Arrays;
-import java.util.List;
-
-import static ru.praktikum.services.qa.scooter.constants.Colors.COLOR_BLACK;
-import static ru.praktikum.services.qa.scooter.constants.Colors.COLOR_GREY;
 import static ru.praktikum.services.qa.scooter.constants.Uri.QA_SQOOTER_PRACTICUM_PROD;
-import static ru.praktikum.services.qa.scooter.steps.StepsCourier.*;
+import static ru.praktikum.services.qa.scooter.steps.StepsCourier.compareMessageFromResponse;
 import static ru.praktikum.services.qa.scooter.steps.StepsOrder.*;
 import static ru.praktikum.services.qa.scooter.utils.Utils.randomString;
 
-@RunWith(Parameterized.class)
-public class OrderCreateTest {
+public class OrdersGetTest extends BaseTest {
 
     private static final String RANDOM_FIRSTNAME_OF_CUSTOMER = randomString();
     private static final String RANDOM_LASTNAME_OF_CUSTOMER = randomString();
@@ -34,30 +27,9 @@ public class OrderCreateTest {
 
     Order order;
 
-
-    @Parameterized.Parameter
-    public List <String> color;
-
-    @Parameterized.Parameters(name = "{index} Сценарий заказа самоката с параметрами с цветом: {0}")
-    public static Object[][] data() {
-        return new Object[][] {
-                { Arrays.asList(COLOR_BLACK) },
-                { Arrays.asList(COLOR_GREY) },
-                { null },
-                { Arrays.asList(COLOR_GREY, COLOR_BLACK) },
-        };
-    }
-
     @Before
     public void setUp() {
         RestAssured.baseURI= QA_SQOOTER_PRACTICUM_PROD;
-    }
-
-
-    @Test
-    @DisplayName("Создание заказа /api/v1/orders") // имя теста
-    @Description("Создаение заказа на основе данных") // описание теста
-    public void postCreateOrder() {
         order = new Order(
                 RANDOM_FIRSTNAME_OF_CUSTOMER,
                 RANDOM_LASTNAME_OF_CUSTOMER,
@@ -67,13 +39,18 @@ public class OrderCreateTest {
                 RENT_TIME_OF_CUSTOMER,
                 DELIVERY_DATE_OF_CUSTOMER,
                 RANDOM_COMMENT_OF_CUSTOMER,
-                color);
-
-        Response responseFromCreateOrder = sendPostRequestToCreateOrder(order);
-        compareMessageFromResponse(responseFromCreateOrder, 201, "{track=");
-
-        order.setTrack(getTrackOrderFromResponseCreate(responseFromCreateOrder));
-
+                null);
+        sendPostRequestToCreateOrder(order);
+        sendPostRequestToCreateOrder(order);
     }
 
+    @Test
+    @DisplayName("Получение доступных заказов /api/v1/orders") // имя теста
+    @Description("Получение двух заказов, доступных для взятия") // описание теста
+    public void getAvailableOrders() {
+
+        Response responseFromGetAvailableOrders = sendGetAvailableOrdersForCouriers(2, 0);
+        compareMessageFromResponse(responseFromGetAvailableOrders, 200, "{orders=");
+
+    }
 }
